Fix `server` to allow periods in querystring
The history-support middleware determines whether a static asset or a page route is requested by using `path.extname` to see if the requested path has a file extension or not. But this results in false positives for querystrings that contain periods.

I added `url.parse(req.url).pathname` to extract just the path of the url, excluding the querystring, and then using `path.extname` on that to check for a file extension.