Issue #66 - Create directory as specified by user input
* Behavior change
* Adds acceptance test to cover new behavior

Currently, running ```ember new FooApp``` will generate a folder called ```foo-app```. As discussed this is a potential source of confusion for users, who will expect a directory called ```FooApp``` to be created.

This commit changes the behavior so that the created directory is called ```FooApp```. However, it maintains the lowercase, dasherized name in the generated package.json file.