Remove RSVP.on('error', ...). Fixes #154
Broccoli's API returns a promise that is rejected when there is a build
error. It is perfectly acceptable to leave this rejection unhandled, as
Broccoli will merrily continue watching and serve a helpful error page
in the meantime. The rejected promise only serves as a hook for API
consumers (like ember-cli) to print out stuff or take other actions on
build errors if they wish.