Fix for ember server --port
Fixes #208

This commit introduces `rewire`, DI library for node.

In order to test `server` command, we need to stub out the
[adapter](https://github.com/stefanpenner/ember-cli/blob/master/lib/commands/server.js#L36).
But because we're testing `server` command in isolation, there's no way
for us to stub it out from outside (pass a stubbed value to the command
itself for example). Hence the DI.