Revert previous CSS preprocessing change
It turns out that may not have been necessary. It seems passing a tree
into the two broccoli sass plugins causes an error, and the static
compiler can still take an array. For whatever reason it's not tracking
this as an extra tree anymore, so it was probably an environment issue.