Indicate that the build was successful after a build error.
Prior to this change if a build errored it was not possible to know from
the console if that issue has been corrected. This forced us to stop &&
restart the server to get feedback (even though it had likely already
been fixed).

This change adds a "Build successful." message when the build had
previously errored (the stack was already printed to the screen), and is
now built properly.

@danmcclain / @rjackson