Fix module location for JSHint tests.
Previously, these tests were output in the same location as the original
file with the `.jshint` extension added on. This causes a number of
problems for anything that assumes a given module path contains certain
types. Specifically, adding an initializer would cause the build to
error since every module in `my-app/initializers` is assumed to export
an object.