Update 2014-04-03-getting-started.md
Removes unnecessary URL associated with http://localhost:4200 as in the rest of the document similar URLs are highlighted and not linked to anywhere. So, I think its good to be consistent. 

Adds http://localhost:4200/tests. As that URL also works out of the box from ember-cli, I think it should be exposed earlier.