Handle flags on $EDITOR
This is for editors like sublime and text mate that will by default
launch and resume execution of io.  In these cases it is usually
reccomended to use the `-w` flag and to set those flags on the $EDITOR
variable in your rc file.

e.g

$EDITOR=mate becomes $EDITOR='/usr/local/bin/mate -w'

This updates the childProcess spawn to trim arguments and pass them as
appropriate.

In the case of mate -w

editor is set to `mate`
editorArgs is set to ['-w']