Add custom static compiler to handle errors.
This is being fixed upstream in
https://github.com/joliss/broccoli-static-compiler/pull/8, but that PR
will not be usable from Ember CLI until the symlink refactor is
completed.

The error swallowing is needed because `multiglob` throws an error if no
files were matched. This becomes a problem for Ember CLI because it is
quite possible that you would not have both pods based templates, AND
non-pods based templates (so one of the `pickFiles` calls would be
empty).