Ensure non `/assets/` files can be served.
Previously, any files in `public/` (outside of `public/assets/`) would
not be served up as it was assumed to match history API requests and
redirect to `/index.html`.

This commit fixes that, by serving up any files that exist in the built
output directory before falling back to checking for history API support
(and subsequently proxying).