Clear require cache for config reloads.
Wrapping in a `Function` worked for simple `config/environment.js`
contents but breaks down in other scenarios that relied on the file
being evaluated in "nodeland".  The prior implementation prevented
custom `require`'s (to get the package name from the `package.json`
for instance).

Now we just clear the `require` cache to allow re-requiring the file
again.