Keep bower_components and vendor separate.
This ensures that `import Baz from 'bower_components/foo-bar/baz';`
works properly. Prior to this change you would have had to know and
understand that we are internally merging `bower_components/` and
`vendor/` together (and therefore you would need to use `import Baz from
'vendor/foo-bar/baz';`).

No fundamental functionality was lost, but it might be a breaking change
for folks that were importing things that happened to be in
`bower_components/`, but were using `vendor/` in the import.  I'll make
sure to call it out as a breaking change in the changelog.