Add some documentation about the .ember-cli file.
It took a fair deal of googling to find out that this configuration file is now available. I think it would be very helpful to add mention of it and how to use it directly in the documentation.

I chose to place this information within this file because it's *sort of* related to a developer's local environment and workflow... but feel it would be great for someone from the Ember-CLI core maintainers to vet both the phrasing and its location.

Thanks for the wonderful tool you've created!