[windows] filterpattern test works on windows
This is the simplest fix to make the test pass. It raises the question of
wether the LiveReloadServer should normalize the slashes in the
filterPatterns so the same pattern can work on different OS. If not the
user would have to craft regex which aren't OS specific, or provided to
filters.