Downgrade to loader.js v1.0.1.
We currently must use 1.0.1 of the loader, due to an inconsistency in
ember-qunit's output.  The 1.x branch of loader.js specifically handles
this issue, but that was removed for 2.0.0.

ember-qunit has the following modules (abbreviated):

```javascript
define('ember-qunit', ['./isolated-container'], function() {});
define('ember-qunit/isolated-container', [], function() {});
```

Clearly `isolated-container` is not in the "same level" as the
`ember-qunit` module.

----

The build system for ember-qunit is actively being rewritten and should
resolve this problem.