Ensure all addon hooks are executed in addon test harness.
Previously, when using the `tests/dummy/` app in an addon project, the
only hooks that were called were build related (because we only added
the addon project to the `addons` list in the `Brocfile.js`).

Now, we always add the addon project. This allows testing commands,
blueprints, middleware hooks, etc.