Adds default description to addon
This change will provide a default description to all addons. This will help prevent addons with missing descriptions which is really commong for ember addons in npmjs.org and emberaddons.com. When a module has no description npmjs.org will try to use the README.md which contains markdown and looks like garbage on both npmjs.org and emberaddons.com.