Update upgrade section of docs
Following up on #2241, this updates the upgrade section of the docs to both point to the release notes and include the steps that are generally suggested with each release.  I think the duplication is worth it in this case because it keeps the upgrade steps more self-contained with the main set of ember-cli docs, but if this proves problematic we could just keep the link to the release notes.