Fix style preprocessors for included addons
Style preprocessors will automatically search for files in addons that
contain a /styles directory, and fail if the addon does not contain
the corresponding file. This ensures that addons will only use
preprocessors if they are required in the dependencies of the addon.

It also fixes the behavior of the getExt() function, which would not
work in addons as they do not pass a relative path the the /styles
directory, and fixes the case where preprocessCSS() is called without
passing an outputPaths option which is also not passed by addons.