Explain Azure Deployment
This commit adds a short explanation of `ember-cli-azure-deploy`, a
small Node module we wrote to make it dramatically easier for Ember Cli
users to deploy their apps to Azure Websites. It essentially enables
automated builds in the cloud, allowing Azure to recompile every single
time a commit is pushed to the project’s repo. Made with :heart: by
Microsoft.