[ENHANCEMENT] use EMBER_NODE_PATH to specify node_modules path
ember-cli will use the value in env.EMBER_NODE_PATH if it is provided
to look up addons when `require`-ing them. This allows a user to
have their `node_modules` path outside of the project root.

Previously, the ember-cli custom require code assumed that
`node_modules` is always in the project's root.