Throw if templating a file fails
Previously, if an error was thrown during the template rendering
process, the error was simply swallowed. This was likely to avoid
errors when attempting to template a binary file, since the byte
stream, when converted to a string, could contain template markers.

Adds the isbinaryfile package to detect if the file is binary. If
so, it skips templating entirely. If not, it attempts to template
the file, and allows exceptions to reject the render promise.

This happened to expose some tests that should have been failing
around the http-proxy generators (the tests were not supplying
enough arguments). Fixes those tests too.