[ENHANCEMENT] Add `nodeModulesPath` property to Project and Addon
This uses the cli utility helper in node-modules-path to
determine the nodeModulesPath.

Also modifies private method `AddonDiscovery.discoverFromDependencies`
to get the nodeModules path from the project (or addon if inside
AddonDiscovery.discoverChildAddons) as its second argument.

This property is intended to provide a way for 3rd-party addons (like
ember-cli-dependency-checker) to know the project's node modules
path without having to discover it in an ad hoc way.