Allow blueprints to omit entity names
Change blueprints to allow for missing entity names. Entity names are
required by default on all blueprints except for an addon's
defaultBlueprint (which will never have an entity name).

This setting can be modified in a blueprint's definition via the
`entityNameRequired` flag, or at runtime by passing the
`entityNameRequired` flag on the options hash to the blueprint's install
method. The latter approach is how defaultBlueprints are exempted
automatically.

This should get rid of the need for empty `normalizeEntityName` hooks
such as
https://github.com/poetic/ember-cli-cordova/blob/master/blueprints/cordova-starter-kit/index.js#L9