Updating startApp to return an instance and the application from separate functions.
Adds an option flag to acceptance test generator for instance vs application setup and teardown

Adding qunit-once hooks to acceptance test generator to request, and destroy, the
application and instance separately.  The application is created and destroyed once
per module, while the instance is created and destoryed for each test.

Updating to passing tests