Remove wasted work in app.import.
In the early days of ember-cli the watcher was very expensive and
we did not want to watch all of `bower_components`, but we have always
intended to allow `app.import`ing a file from a given directory to
allow that directory to be watched.

The way that we selectively watched these subdirectories in
`bower_components` was by creating a tree for each file `app.import`'ed
and then merging those together but not using any of their output!! This
basically meant that for every `app.import` we created an extra
ultimatley unused `Funnel` instance that we merged with the rest of the
system resulting in merging some subdirectories of `bower_components`
with themselves AND the rest of `bower_components` multiple times over
again in each build.

This change removes all that crazy shenanigans and if `watchman` is
present, just watches all of `bower_components`.