remove confusing debug logging.
outputDir has nothing to do with this, as the files are served
from the latest broccoli build directory not `dist`.

This is the case, to allow concurrent broccoli builds of different environments.
If we served from `dist` we would serve w/e the last build was, regardless
if it matched the environment the server is running in