Support custom blueprint options in new and init commands
The new and init commands can be given a custom blueprint to use. This modifies those commands to be aware of blueprint-specific options and pass them through, like the generate command.

This is handy in the near-term for custom app/addon blueprints that accept options, and could pave the way to add options to the build-in blueprints if that ever seems valuable. A few examples I can think of immediately are the server port and the root URL.