print double dash for long aliases in availableOptions
note: presently if one alias for the availableOptions is a long word
the help function for that command will still print the option with a single
dash although the `--` version seems to work fine

i.e.
https://github.com/ember-cli/ember-cli/blob/282641ba662292afb40cb88cdf31557a3e4cc6b7/lib%2Fcommands%2Fbuild.js#L13
will be printed as `-dev` even if also `--dev` is supported.

before:
aliases: -a <value>, -long-a <value>, -b (--test-option=c)

after:
aliases: -a <value>, --long-a <value>, -b (--test-option=c)