Enable X-Forwarded-* headers in http-proxy
Reasoning: when using e.g. a Rails backend running on a separate port, proxying it using the ember-cli proxy feature, the generated _links_ sometimes tends to be incorrect. They are based on settings in the X-Forwarded-* headers. Hence, xfwd must be enabled.

If you like, we can make this a command-line option.

/cc @perlun, @jesjos