Remove require cache caused shared mutation in slow tests.
Prior to this change, the require cache was causing us to mutate a
single shared `package.json` (one for `some-cool-app` and one for
`some-cool-addon`), but each of these tests assume they are modifying
the default blueprints `package.json`. Instead, they were all mutating a
shared copy (cached on the first test).