Check for `ember-app` or `ember-addon` when looking for `package.json`
In the event that a package.json file somehow ends up above your current
working directory, ember-cli would previously become very confused
because it assumed that the directory with the package.json file was
the root of an ember-cli app or addon. This commit updates the search
algorithm to verify that the package.json file found is indeed an
ember-cli app or addon, based on the package.json keywords.

A joint production of @ryanlabouve and @lukemelia, began at EmberConf
2016 Contributors Workshop.

Fixes #5493