models/addon: Kill the addonJsFiles() cache
addonJsFiles() is called twice, by jshintAddonTree() and by treeForAddon(). in one case it gets passed the addonPath directly, while the other case might be overwritten and passes something else. If the Funnel in the first invocation is cached the second invocation parameter will be discarded which resulted in a regression.