Uniform import API
This PR proposes a uniform API for importing assets into both apps and addons. It solves the most common pain point when using an addon from another addon, which is that `app.import` in the `included` hook fails.

This API gives us future flexibility to implement `import` differently depending on target you're importing into. In particular, an engine may choose to behave differently  so that vendored assets are in the appropriate bundles.

I'm opening this for discussion now, and will add tests if people think this is mergeable.