Allow custom config paths
If you are running an Ember App not in the root directory of your app, you need to be able to specify where your config is located in order for Ember CLI to run correctly.
This adds the option to allow configuring the configPath via package.json - specifically via package.directories.configPath.