[BUGFIX beta] Ensure `--no-color` option is present for all commands.
Using this option has always worked to disable `chalk` from actually
emitting colored output (and therefore already has the desired effect),
but it would emit a warning to the user when used suggesting that the
option was not valid.

This adds `--no-color` as a global option to all commands, and adds
a test to ensure that no warning is printed when it is used.