Prevent deprecation from `ember-cli-babel` config options.
Future versions of ember-cli-babel will be moving the location for its
own configuration options out of `babel` and will be issuing a deprecation
if used in the older way.

This changes things around to detect the correct (undeprecated) options
location, and use it when possible.  Otherwise it falls back to `babel`
(which is backwards compatible without a deprecation and forwards compatible
with a deprecation).