Allow using empty directories.
The usecase of `--directory` inside of Ember CLI allows you to specify where
you wish to place the output from the `ember new` and `ember addon` commands.
It's likely in those scenarios that the directory you're trying to populate
already exists. Our goal was to make sure we played nicely with the existing
assets. An empty directory meets that bar for safety and improves the
ergonomics of the `--directory` command, also making it more suitable for
testing.