Introduce command interruption
Fixes the problem when the Builder always exits with an exit code 1 when
the process is interrupted.

This pr moves making of decision about an exit code from the builder to
the tasks.

cli now listens to the interruption signals(ctrl+c) and tries to interrupt the current
active command gracefully which means interruptin of the current active
task.