Remove wasted work around addon's addon trees.
Prior to this change, after retrieving the `addon` trees for all addons
we then immediately concat the assets down into a single `addons.{js,css}`
file (after merging and funneling to a different location).

The long term objective is to keep "loose modules" throughout the entire
pipeline and eventually replace the final concat stages that we are currently
using with a higher level "packager". This takes a very small step towards
that future...

The good news is that this is actually a win regardless of that future state.