Ensure all blueprints are running in strict mode.
Strict mode is required for using `const` / `let` in module scope under Node 4.

We had previously disabled strict mode requirement for `blueprints/` directory,
due to the fact that the majority of files were actually intended for the browser
and are already receiving a global use strict.

This removes the `blueprints/.eslintrc.js` override and fixes the resulting errors
regarding missing `'use strict';` in each of the blueprint files.