Preserve user errors in instrumentation hook
When we run `stopAndReport('init')` we invoke user hooks.  These might
error.  If they do the instrumentation start/stops will be unbalanced
and we will report the following unhelpful error instead of the actual
error in the user hook.

`Error: Cannot stop instrumentation "command".  It has not started.`

This commit fixes this issue.