Move instrumentation out of `bin/ember`.
As has generally been our preference, we try very very hard to avoid
changing things in `bin/ember` that may negatively impact users using
it as a launcher for many different ember-cli versions.

This moves the recent changes out of `bin/ember` into `lib/cli/index.js`.
The downside is that it is _slightly_ later in the boot process, but
the upside is that we get to keep `bin/ember` backwards and forwards
compatible for a little while longer...