Remove guard in `treeForAddon` around `addon/**/*.js` files.
This guard was added to avoid wasted effort when preprocessing
the `addon/` directory when it only contains a single `.gitkeep`
file (or is non-existent). The assumption I made was that if
there are no `*.js` files inside the `addon/` directory, then
we didn't need to attempt to transpile it.

Unfortunately, "back in reality" there are a number of addons
that wrap existing libraries, that have no `addon/**/*.js` files
but still need the `this._super.treeForAddon` invocation to process
the provided tree.

This change, reverts the guard (it was only a relatively small savings
anyways).