Watch-detector clean up
Addresses #6311.

Refactors the watch detector code to be clearer to read. The logic
used to detect node/watchman/etc have not been changed much, they've
just been updated to not include any fallback logic and to not write
and text back to the UI.

The basic logic of selecting a watcher is to first try the user's
preference, if provided, otherwise watchman, node, and polling
watchers are tried, in order. The `events` type isn't documented
anywhere, AFAICT, so it might be a candidate for removal since the
default order covers the events type anyway.