Update ISSUE_TEMPLATE.md
Now that ember-cli supports using yarn, it is likely that more issues will come through with regards to yarn usage and integration with ember-cli. Because of that, we should ask users to include their yarn version number in issue reports. If they are not using yarn, we'll simply see a "command not found" message as the last line in the output.