Ensure `lintTree` results cannot clobber tests.
The prior logic resulted in the result of `lintTree` being
**after** the actual processed `tests/` tree. When those trees
were then passed to `MergeTrees` the `lintTree` result would
clobber the normal processed `tests/` tree. In general, when
a linter is actually emitting linting tests, this is fine
because the files that it emits are not the same names as the
original (e.g. `tests/test-helper.js` becomes
`tests/test-helper.lint-test.js`), but when the `lintTree`
implementation does not emit different names than the input tree
the `lintTree` versions "win".

In Ember CLI < 2.13, this setup still worked properly because we
would fully process the result of `lintTree` through the consuming
applications JS processor pipeline (e.g. babel). This _sounds_ neat
but actually means that we were transpiling all of `tests/` twice.

In Ember CLI 2.13 and above, we only process the result of `lintTree`
for modules (not a full "transpile down to ES5"). This means that
any babel processing that would have been done for `tests/` normally
is thrown away when the transpiled versions are clobbered by the
untranspiled `lintTree`.

This change corrects the ordering of trees, so that the correctly
proccessed `tests/` tree "wins" when the `lintTree` tree emits files
with the same names.