[INTERNAL] Deprecate `bundler.js`
This is a non-breaking, backwards compatible change.

After talking to Rob, we decided it might be better to
inline `broccoli-assembler` for now. That way it'll give
a chance for people to get familiar with what it does
and extract it later into a separate package under `ember-cli` org.

The reason behind deprecation is that we do not want to
have both names in the future: "bundler" and "assembler".

Changes:

+ remove `bundler` and associated tests
+ add `assembler` and tests
+ add `strategies` to encapsulate concatenate logic for JS
+ refactor `ember-app` to accomodate changes