[EXPERIMENT] Standard ESM
This is experimental commit based on the on of the conversations we had during Ember CLI core meeting. This change allows us to start using `import`/`export` across the codebase.

This change enables a smooth transition between Node and ES module formats and compensates for CJS and ESM incompatibility.

The tool used to enable this transition is [@std/esm](https://github.com/standard-things/esm).

`@std/esm` uses [reify](https://github.com/benjamn/reify). The `@std/esm` loader is intended for runtime use, so no build step or pre-publish step is required of the user. There is a 1 time transform cost per ESM module, after that the result is cached and loading is on par with CJS. The loader itself when published will be a small, ~40kb, zero dependency package.

Changes:

+ add `@std/esm` as a dependency
+ add `@std/esm` config to `package.json`
  + `"cjs": true` means we get to use `require` in files until we’re ready to convert to `import`
  + `"esm": "js”` means we don’t have to convert file extensions from `*.js` to `*.mjs`
+ repurpose `lib/cli/index.js` to use `@std/esm` loader
+ move `lib/cli/index.js` code to `lib/cli/main.js` and replaced `require`s with `import`s
+ remove `use strict;` from files as `@std/esm` adds it by default
+ add `overrides` to `.eslintrc.js` to hint different parse option
+ add a note on ESM support

Relevant links:

+ [ES Modules in Node Today!](https://medium.com/web-on-the-edge/es-modules-in-node-today-32cff914e4b)

Caveats:

_Usage of ‘use strict’;_. If we decide to go down the road of conversion, we should drop the usage of `use strict` not only in Ember CLI internally but in the blueprints as well (`ember-cli-build.js`, `config/environment.js`, `targets.js`).