Fix issue with linting within an addon without an `app/` directory.
ember-cli 2.16 added support for linting the `app/` tree of an addon
(while developing the addon itself). Unfortunately, we did not guard
generation of the linting tree based on the presence of `app/` (addons
are not required to have any specific directories).

This ensures that we only invoke `lintTree` on `app/` if it is present.