Remove ember-cli-legacy-blueprints.
ember-cli has supported consuming `ember-source` for quite a number of
versions now (since 2.12), the `ember-cli-legacy-blueprints` addon is
only supporting usage of blueprints with versions of Ember prior to
2.12.

This commit removes `ember-cli-legacy-blueprints` from a default new
project. If a given project still needs this addon, it should add it as
its own dependency. Once added to the projects own dependencies the
legacy blueprints will be available exactly the same as today.