Give plugins and extends their own lines
I had several extra `plugins` and `extends` entries defined in some of my `.eslintrc.js` files. The idea here is, if we put each one on a separate line, and you use the default suggestions, plus add your own, you won't get a diff where the whole plugins and extends sections are completely rewritten, but the diff will just show the extra ones you added. Please let me know if you guys think this is a good idea or not!