Use bower from project root instead of relative to file path
During testing we change the current directory via `process.chdir(tmpdir)`. During
tests that run `ember new` this tmp directory becomes the project.root. The
bower executable should be installed into tmp/X/node_modules and not
ember-cli's node_modules. The error this was causing is that if bower
was not found (and it would not be as ember-cli does not depend on bower) bower would
install itself into ember-clis package.json and corrupt that file. This
now correctly uses the project.root as to where it should look for
bower.