Extract `contentFor` into `ember-app-utils`
Main motivation behind this change is on-going extraction into default
packager. Because, we previously bound `contentFor` to application's
context, it was harder to extract the way we assemble certain trees
(config tree, for example).

This change does the following:

+ removes that binding and replaces it with explicit contract (we pass
now pass in additional `options` object) which makes it easier to test
(no need to set up `EmberApp`) and easier to reason about
+ adds missing tests for both "app-boot" and "test-body-footer"
types
+ removes `contentFor` method from `EmberApp`: not a breaking change
since it is a private method