Make sure addon transforms are available in app
I noticed when using ember-cli-cjs-transform in one of our addons, that when it was consumed in an app, the transform was flagged as unavailable. I believe this should fix the issue. Most of this idea was given to me by @rwjblue. Please let me know if it needs any tweaks!