Bring back concept of default plugin for type.
Instead of using the previous heuristics we require that a plugin
declare itself as the default for type. This would look like:

```js
registry.add('js', {
  name: 'ember-cli-babel',
  ext: 'js',
  isDefaultForType: true,
  toTree: (tree) => this.transpileTree(tree)
});
```

If no plugins declare that they are the default for a given type, all
plugins will be used (since choosing the default arbitrarily is not a
valid option).