Ensure `treeForSrc` does nothing when `isModuleUnification` is falsey.
An earlier change made `isModuleUnification` a manual opt-in (due to the
breaking change of detecting `src/` and assuming it was module
unification), however the `treeForSrc` implementation was not properly
leveraging `this.isModuleUnification()` to determine if it should even
do any work.

This adds an additional conditional to the short curcuit, and ensures
that `treeForSrc` does _nothing_ when a given addon is not opting in to
module unification.