Do not attempt to compress server sent events.
From the `compression` packages README:

> Because of the nature of compression this module does not work out of
> the box with server-sent events. To compress content, a window of the
> output needs to be buffered up in order to get good compression.
> Typically when using server-sent events, there are certain block of data
> that need to reach the client.

There are work arounds to this (every middleware that handles SSE's
would have to immediately call `res.flush()`), but they do not work when
proxying (because it is impossible for the proxy itself to know when
`.flush()` should be called.

This change updates the compression code to simply disable itself for
`text/event-stream` (which is what SSE uses as its mime type).