Remove `appAndDependencies` method
All the processing logic for application files (javascript and
templates) and dependencies (import transforms) has moved to the
packager. That way `ember-app` is only responsible for providing
application and add-on files that are needed.

Because of that `appAndDependencies` method is no longer needed and
could be removed altogether. It's a private method so it is safe to do
so.