Config hook called too soon during serve
The config hook gets called eagerly during the serve flow and that
leads to issues with the context of the config hooks not being set
up properly. An example of this is if you have:

config() {
  return this.app && this.app.options[this.name];
}

`this.app` will be undefined but that same hook during a build will
not be undefined.