Explicitly include `this.trees.styles`
After packager work landed, there was no need to explicitly merge
`app.trees.styles` into a "full tree" as for classic model they would
be a part of "app" tree, for MU they would be a part of "src" tree.

As it turned out some developers relied on the fact that they could add
things to `app.trees.styles` so that packager changes turned out to be
breaking changes.

Since this behaviour manifested only when using MU *and* adding styles
to `app.trees.styles`, this is not critical.