Turn experiments object into a function
When testing an experiment in `ember-cli` scope, a new CI scenario seems
to be the go-to solution. This involves scopping tests like this:
```js
if (experiments.MY_EXPERIMENT) {
  describe('my experiment module', ...)
  // or
  module('my experiment module', ...)
} else {
  describe('regular module', ...)
  // or
  module('regular module', ...)
}
```

But when testing an experiment outside of `ember-cli` scope, things get
different. And we may want to be able to do:

```js
describe('my addon blueprint', {
  it('works - in a classic app', ...);

  it('works - in a MU app', ...);
})
```

The issue is: as an object, `experiments` is cached between the two
tests (run on the same process). In these scenarios, we need not to
cache the object, hence the function.