Stabilize + Lock-down add-on discovery order.
The recent PackageInfoCache work accidentally re-ordered dependencies, applying devDependencies after dependencies rather then before. This resulted hard to diagnose but breaking behavior.

An interim solution, sorted dependencies wholesale lexicographically but this resulted in other breakage.

To address all this, this PR:

* restores the original precedence (devDependencies < dependencies)
* ensures ordering within devDependencies or dependencies is always lexicographical (regardless of the order within package.json)
* if duplicates exist between devDependencies and dependencies, we take a “element of least surprise approach” they are de-duped, and the “last occurrence” encountered solidifies the ordering.  order([a,b,c,a]) => [b,c,a]
* thorough testing to lock down the ordering, ensuring future work

This aims to remove preventable ambiguity. For those requiring explicit ordering beyond what we support out of the, they are encourage to use the before/after API. This will create an explicit edge, forcing order…

This PR was tested against a very large real world application. This application has 40+ engines, and a huge number of add-ons. This application was also the one which detected the above issues. After this PR everything works!