Align PackageInfoCache platform specific path usage with the rest of the code-base.
In general, ember-cli avoids unnecessary path normalization.
Typically, passing user input un-normalized to fs methods, the exceptions
are when we perform string mutation or path comparisons.

The goal here, is to ensure all API’s can handle both posix and win32 paths,
and that internal data-structures are consistent regardless of which platform
they are run on.

Related #8006 #8002 #7998