Ensure tiny-lr server is properly setup.
We are not calling `tiny-lr` servers `configure` hook (which is normally
how it would be created) and are instead setting things up ourselves.
Unfortunately, when `close` is called tiny-lr checks its `.server` to
ensure it is also closed (in case it started its own).

This change sets the `this.server` property on the created `tiny-lr`
Server to avoid errors during teardown...