Reset blueprint model's ignoredFiles
When running `ember new` or `ember addon` in sequence, on the same node
process, `ignoredFiles` gets properly incremented. But it's never reset.

This change is especially usefull in tests.

This PR aims at reseting the key so that the following test setup
works properly:
```javascript
beforeEach(() => {
  emberNew({ isModuleUnification: true, target: 'addon' })
})
```

The bug occured while testing the `addon` blueprint for the Module
Unification Quest. Because in the tests, we want to make sure the `src`
directory exists. Previously, `addon` directory was disapearing but
the issue remained silent as we were not asserting its existence.