DefaultPackager: Move `addon-test-support` out of the `tests` folder
`packageApplicationTests()` packages *all* files in the `tests` folder, which caused the `addon-test-support` files to be packaged in the `tests.js` *and* the `test-support.js` files. By moving `addon-test-support` out of the `tests` folder for the downstream trees, we can make sure that `packageApplicationTests()` only packages actual test files and `packageTestFiles()` correctly picks up the `addon-test-support` files.