DefaultPackager: Fix addon preprocessing
`preprocessTree('test', ...)` assumes that the input tree has the tests in an `<app-name>/tests` folder, but the default packager implementation regressed and passed the tests in an non-nested `tests` folder. This commit fixes the packager to correctly pass the tests in `<app-name>/tests` again, by moving the `callAddonsPreprocessTreeHook()` call below the `Funnel` that does the `<app-name>` prefixing.