Remove last usage of Babel 6.
This last usage is _only_ used for supporting compiling templates in
`addon/` and supporting `addon-test-support/` in addons that themselves
do not have any `.js` processors (e.g. no dependency on
`ember-cli-babel`). Note: those scenarios have been emitting a warning
for quite some time...

This change is *not* a breaking change to our users (the module
compilation between Babel 6 and Babel 7 is compatible).