Refactor template build pipeline to enable co-located templates.
This refactors the processing of templates so that instead of filtering out
templates from JavaScript we process them in a pipeline (first the templates,
and the result of that is passed in to be compiled as JavaScript).

On its own, this change does not enable co-located templates, but this work is
required so that ember-cli-htmlbars can implement the feature.

Co-authored-by: Robert Jackson <me@rwjblue.com>