Avoid using configstore when analytics is disabled
configstore crashes if the user's $HOME directory is not writable.

Usually a user's $HOME directory is writeable, but the [Nix][nix]
build system purposefully points $HOME at a non-existent directory
to ensure an isolated, repeatable build environment.

This creates a workaround so that disabling analytics will not
trigger the crash