[BUGFIX] Don't reload the config for instrumentation
We currently will reload the config to setup instrumentation, which can
happen multiple times while the CLI is running. This sometimes results
in failures when the CLI generates new applications that haven't
installed `node_modules` yet, and in other cases. We really only need to
load the config once, so this PR changes it so that we do.