Update GitHub Actions setup to use TAP output.
The other CI systems already use TAP output, but GitHub Actions didn't
(because we don't set `CI` environment variable in our config, and they
don't set it automatically).

Using TAP output in CI makes it possible to Cmd+F (or Ctrl+F) for "not
ok" to jump directly to the failing tests, otherwise with the standard
Mocha `spec` reporter it is quite difficult to jump to the failures.