Fix issue in `treeForSrc` for addon's with `src` but no template compilers.
We should _probably_ remove the module unification build support, but
that is 1) more work and 2) breaking to folks still using module
unification. The goal is to remove that _after_ 3.16 is released (so
they'll have seen the deprecation warnings in the LTS version).

This broke due to `@ember-data/private-build-infra` using `src/` folder
(for non-MU things). In `@ember-data/private-build-infra@3.15.0` that
package was **not** an addon (so this issue didn't crop up), but it was
changed to be an addon in 3.16.0.